feat: introduce jwt-spi and jwt-core modules
