+++ CODEOWNERS
@@ -25,8 +25,6 @@ styleguide.md @paullatzelsperger
 /buildSrc/ @paullatzelsperger @marcgs
 
 /common/state-machine-lib/ @ndr-brt @algattik
-/common/token-generation-lib/ @bscholtes1A
-/common/token-validation-lib/ @bscholtes1A
 
 /core/base/agent/ @jimmarino
 /core/base/policy/ @jimmarino
