+++ common/util/build.gradle.kts
@@ -35,7 +35,6 @@ publishing {
         create<MavenPublication>("common-util") {
             artifactId = "common-util"
             from(components["java"])
-            artifacts.forEach { a -> println(a.file) }
         }
     }
 }
