+++ docs/developer/modules.md
@@ -1,7 +1,5 @@
 
 org.eclipse.dataspaceconnector:state-machine-lib:0.0.1-SNAPSHOT
-org.eclipse.dataspaceconnector:token-generation-lib:0.0.1-SNAPSHOT
-org.eclipse.dataspaceconnector:token-validation-lib:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:common-util:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:ids:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:data-plane-server:0.0.1-SNAPSHOT
@@ -9,6 +7,7 @@ org.eclipse.dataspaceconnector:module-domain:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:module-processor:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:core-base:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:core-boot:0.0.1-SNAPSHOT
+org.eclipse.dataspaceconnector:jwt-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:policy-engine:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:policy-evaluator:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:contract:0.0.1-SNAPSHOT
@@ -42,6 +41,7 @@ org.eclipse.dataspaceconnector:auth-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:catalog-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:core-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:identity-did-spi:0.0.1-SNAPSHOT
+org.eclipse.dataspaceconnector:jwt-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:oauth2-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:transaction-datasource-spi:0.0.1-SNAPSHOT
 org.eclipse.dataspaceconnector:transaction-spi:0.0.1-SNAPSHOT
