+++ extensions/common/iam/oauth2/daps/src/main/java/org/eclipse/dataspaceconnector/iam/daps/DapsJwtDecorator.java
@@ -16,7 +16,7 @@ package org.eclipse.dataspaceconnector.iam.daps;
 
 import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jwt.JWTClaimsSet;
-import org.eclipse.dataspaceconnector.common.token.JwtDecorator;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecorator;
 
 public class DapsJwtDecorator implements JwtDecorator {
     @Override
