+++ extensions/common/iam/oauth2/oauth2-core/build.gradle.kts
@@ -22,8 +22,7 @@ val okHttpVersion: String by project
 
 dependencies {
     api(project(":spi:common:oauth2-spi"))
-    implementation(project(":common:token-generation-lib"))
-    implementation(project(":common:token-validation-lib"))
+    implementation(project(":core:common:jwt-core"))
 
     implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
     implementation("com.squareup.okhttp3:okhttp:${okHttpVersion}")
