+++ extensions/common/iam/oauth2/oauth2-core/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/core/identity/Oauth2ServiceImpl.java
@@ -20,16 +20,16 @@ package org.eclipse.dataspaceconnector.iam.oauth2.core.identity;
 import okhttp3.FormBody;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
-import org.eclipse.dataspaceconnector.common.token.JwtDecorator;
-import org.eclipse.dataspaceconnector.common.token.JwtDecoratorRegistry;
-import org.eclipse.dataspaceconnector.common.token.TokenGenerationService;
-import org.eclipse.dataspaceconnector.common.token.TokenValidationService;
 import org.eclipse.dataspaceconnector.iam.oauth2.core.Oauth2Configuration;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 import org.eclipse.dataspaceconnector.spi.iam.IdentityService;
 import org.eclipse.dataspaceconnector.spi.iam.TokenParameters;
 import org.eclipse.dataspaceconnector.spi.iam.TokenRepresentation;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecorator;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecoratorRegistry;
+import org.eclipse.dataspaceconnector.spi.jwt.TokenGenerationService;
+import org.eclipse.dataspaceconnector.spi.jwt.TokenValidationService;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.eclipse.dataspaceconnector.spi.types.TypeManager;
