+++ extensions/common/iam/oauth2/oauth2-core/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/core/jwt/DefaultJwtDecorator.java
@@ -17,7 +17,7 @@ package org.eclipse.dataspaceconnector.iam.oauth2.core.jwt;
 import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jose.util.Base64URL;
 import com.nimbusds.jwt.JWTClaimsSet;
-import org.eclipse.dataspaceconnector.common.token.JwtDecorator;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecorator;
 
 import java.time.Clock;
 import java.util.Date;
