+++ extensions/control-plane/data-plane-transfer/data-plane-transfer-sync/src/main/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/proxy/DataPlaneProxyTokenDecorator.java
@@ -16,7 +16,7 @@ package org.eclipse.dataspaceconnector.transfer.dataplane.sync.proxy;
 
 import com.nimbusds.jose.JWSHeader;
 import com.nimbusds.jwt.JWTClaimsSet;
-import org.eclipse.dataspaceconnector.common.token.JwtDecorator;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecorator;
 
 import java.util.Date;
