+++ extensions/control-plane/data-plane-transfer/data-plane-transfer-sync/src/main/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/validation/ExpirationDateValidationRule.java
@@ -15,7 +15,7 @@
 package org.eclipse.dataspaceconnector.transfer.dataplane.sync.validation;
 
 import com.nimbusds.jwt.SignedJWT;
-import org.eclipse.dataspaceconnector.common.token.TokenValidationRule;
+import org.eclipse.dataspaceconnector.spi.jwt.TokenValidationRule;
 import org.eclipse.dataspaceconnector.spi.result.Result;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
