+++ gradle.properties
@@ -36,7 +36,7 @@ jupiterVersion=5.8.2
 jwtVersion=3.13.0
 micrometerVersion=1.8.2
 mockitoVersion=4.2.0
-nimbusVersion=8.22.1
+nimbusVersion=8.23
 okHttpVersion=4.10.0
 openTelemetryVersion=1.12.0
 postgresVersion=42.4.0
