+++ spi/common/jwt-spi/build.gradle.kts
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+plugins {
+    `java-library`
+    `maven-publish`
+}
+
+val nimbusVersion: String by project
+
+dependencies {
+    api(project(":spi:common:core-spi"))
+
+    implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
+}
+
+publishing {
+    publications {
+        create<MavenPublication>("jwt-spi") {
+            artifactId = "jwt-spi"
+            from(components["java"])
+        }
+    }
+}
