+++ spi/common/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/Oauth2JwtDecoratorRegistry.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.iam.oauth2.spi;
 
-import org.eclipse.dataspaceconnector.common.token.JwtDecoratorRegistry;
+import org.eclipse.dataspaceconnector.spi.jwt.JwtDecoratorRegistry;
 
 /**
  * Registry for Oauth2 jwt decorators.
