+++ spi/common/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/Oauth2ValidationRulesRegistry.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.iam.oauth2.spi;
 
-import org.eclipse.dataspaceconnector.common.token.TokenValidationRulesRegistry;
+import org.eclipse.dataspaceconnector.spi.jwt.TokenValidationRulesRegistry;
 
 /**
  * Registry for Oauth2 validation rules.
